Deprecate APM\Command::getServer(), add getPort and getHost instead (#2773)
CommandSucceededEvent|CommandFailedEvent::getServer() has been deprecated in ext-mongodb 1.20 and removed in 2.0.0.
For backward compatibility: we keep this method in MongoDB ODM, but throw an exception the extension is not compatible. I'm not adding a trigger_deprecation as this is already triggered by the extension method call.

The new methods getPort() and getHost() are added as proxy to extension methods that are always available in the required ext-mongodb 1.21+.

This is not covered by tests, I can add tests on CommandLogger before merging.