Deprecate not overriding `Type::closureToPHP()` (#2960)
The default implementation of Type::closureToPhp() is incorrect as it forwards the unmodified value.
In 3.0, I want to change the default implementation to use the one from the ClosureToPHP trait that calls the type class.
Also adding the missing implementation for collection and hash.
Any type that does not override the closureToPhp will get a deprecation notice.